Refactor: - move streams into distinct modules - Make testing mode as decorator
