+++ airbyte-integrations/connectors/source-mixpanel/source_mixpanel/streams/__init__.py
@@ -1,0 +1,18 @@
+from .annotations import Annotations
+from .cohort_members import CohortMembers
+from .cohorts import Cohorts
+from .engage import Engage
+from .export import Export
+from .funnels import Funnels, FunnelsList
+from .revenue import Revenue
+
+__all__ = [
+    "Engage",
+    "Export",
+    "CohortMembers",
+    "Cohorts",
+    "Annotations",
+    "Funnels",
+    "FunnelsList",
+    "Revenue",
+]
